fix(MarkLine): modify the issue where the thershold was too high
